Made a rounds component
I added a basic start button and how to get the initial points of each player after their double rolls. At the end it tells the user if they are eligible to win double points or not. Same goes for the computer. I added a loop where if both the computer and user are <= 13 then they keep going, it contains a question each loop where it asks if the player wants to roll or pass their turn and after that the computer rolls. At the end of both their turns it displays how much they got from each roll and shows them their updated points.