Fixed doubling of points
I fixed this by making the variables user_double_points and computer_double_points. This was because the double_points variable had been overridden as I used double_points = user_first[1] and then used the same double_points for computer_first. This made it so that the stored information inside double_points before are now gone or rather it has the computer’s information. So when I won and was eligible for double points, I didn’t get any as the computer wasn’t eligible for double points.