Added a tie condition and computer pass
Added a tie condition where if the computer and user have equal points and both sides have passed then it would result in a tie. There was also the addition for the computer's ability to pass where if the computer has more than or equal to 10 points and less than or equals to 13 points while also having >= user_points then it automatically passes.